**READ BELOW - Bug Report**
After letting a few users sign up and use the recipe book. Some had issues with registering for the site.
Resulting in the user seeing a white empty page when trying to sign in with the account created.

After changing the passwords around, it concluded there was an issue with the unicode keychain shuffle I implemented for some basic encyption.

This commit contains the same encrypter but with a smaller param value.